+++ src/main/java/at/blvckbytes/component_markup/markup/ast/tag/built_in/colorize/rainbow/RainbowGenerator.java
@@ -4,7 +4,7 @@ import at.blvckbytes.component_markup.markup.interpreter.PackedColor;
 
 public class RainbowGenerator {
 
-  public int getPackedColor(double progressionPercentage) {
+  public long getPackedColor(double progressionPercentage) {
     float h = (float) ((progressionPercentage / 100) * 6);
 
     int i = (int) Math.floor(h);
