+++ src/main/java/at/blvckbytes/component_markup/markup/ast/tag/built_in/colorize/rainbow/RainbowNodeState.java
@@ -14,7 +14,7 @@ public class RainbowNodeState extends ColorizeNodeState {
   }
 
   @Override
-  protected int getPackedColor(double progressionPercentage) {
+  protected long getPackedColor(double progressionPercentage) {
     return rainbowGenerator.getPackedColor(progressionPercentage);
   }
 }
