+++ src/main/java/at/blvckbytes/component_markup/markup/ast/tag/built_in/colorize/transition/TransitionNodeState.java
@@ -22,7 +22,7 @@ public class TransitionNodeState extends GradientNodeState {
   }
 
   @Override
-  protected int getPackedColor(int index, int length) {
+  protected long getPackedColor(int index, int length) {
     return super.getPackedColor(0, 1);
   }
 }
